Move all emitter and channel tests in a dedicated package
