+++ extensions/smallrye-reactive-messaging/deployment/src/test/java/io/quarkus/smallrye/reactivemessaging/channels/DeprecatedChannelConsumer.java
@@ -1,17 +1,17 @@
-package io.quarkus.smallrye.reactivemessaging;
+package io.quarkus.smallrye.reactivemessaging.channels;
 
 import java.util.List;
 
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
 
-import org.eclipse.microprofile.reactive.messaging.Channel;
 import org.eclipse.microprofile.reactive.messaging.Message;
 
 import io.reactivex.Flowable;
+import io.smallrye.reactive.messaging.annotations.Channel;
 
 @ApplicationScoped
-public class ChannelConsumer {
+public class DeprecatedChannelConsumer {
 
     @Inject
     @Channel("source")
