+++ extensions/smallrye-reactive-messaging/deployment/src/test/java/io/quarkus/smallrye/reactivemessaging/channels/EmitterWithOverflowTest.java
@@ -1,4 +1,4 @@
-package io.quarkus.smallrye.reactivemessaging;
+package io.quarkus.smallrye.reactivemessaging.channels;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -13,15 +13,15 @@ import org.junit.jupiter.api.extension.RegisterExtension;
 
 import io.quarkus.test.QuarkusUnitTest;
 
-public class EmitterWithBroadcastTest {
+public class EmitterWithOverflowTest {
 
     @RegisterExtension
     static final QuarkusUnitTest config = new QuarkusUnitTest()
             .setArchiveProducer(() -> ShrinkWrap.create(JavaArchive.class)
-                    .addClasses(EmitterWithBroadcastExample.class));
+                    .addClasses(ChannelEmitterWithOverflow.class));
 
     @Inject
-    EmitterWithBroadcastExample bean;
+    ChannelEmitterWithOverflow bean;
 
     @Test
     public void testEmitter() {
@@ -32,11 +32,17 @@ public class EmitterWithBroadcastTest {
         assertEquals("b", list.get(1));
         assertEquals("c", list.get(2));
 
-        List<String> list2 = bean.list2();
-        assertEquals(3, list2.size());
-        assertEquals("a", list2.get(0));
-        assertEquals("b", list2.get(1));
-        assertEquals("c", list2.get(2));
+        List<String> sink1 = bean.sink1();
+        assertEquals(3, sink1.size());
+        assertEquals("a1", sink1.get(0));
+        assertEquals("b1", sink1.get(1));
+        assertEquals("c1", sink1.get(2));
+
+        List<String> sink2 = bean.sink2();
+        assertEquals(3, sink2.size());
+        assertEquals("a2", sink2.get(0));
+        assertEquals("b2", sink2.get(1));
+        assertEquals("c2", sink2.get(2));
     }
 
 }
